Add destructor and deep copy constructor to Deep class
A destructor was added to the `Deep` class to ensure proper
deallocation of dynamically allocated memory (`data`) and to
log a message when the destructor is called.

A deep copy constructor was implemented to allocate new memory
for the `data` member and copy the value from the source object,
ensuring safe copying of `Deep` objects. A log message was added
to indicate when the deep copy constructor is invoked.